+++ bpf/insn/samples/bpf_insn_defs.h
@@ -1,4 +1,5 @@
 BPF_INSN_SAMPLE(perf_event_output)
+BPF_INSN_SAMPLE(probe_read_kernel_str)
 BPF_INSN_SAMPLE(trace_printk)
 BPF_INSN_SAMPLE(get_current_cgroup_id)
 BPF_INSN_SAMPLE(get_current_task_btf)