+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 #### Breaking Changes
 #### Added
 - bpftrace: playground: add asm.bt test;
+- bpf: test `bpf_probe_read_kernel_str()`, `linux >= v5.5`;
 - bpf: test `bpf_perf_event_output()`, `linux >= v4.4`;
 - bpftrace: add playground subdir;
 - uki: add makeuki.sh;